+++ apps/example-nest/src/app/post-validate-responses.controller.ts
@@ -39,7 +39,7 @@ export class PostValidateResponsesController
   @TsRest(c.getPosts)
   async getPosts(
     @TsRestRequest()
-    { query: { take, skip, search } }: RequestShapes['getPosts']
+    { query: { take, skip, search } }: RequestShapes['getPosts'],
   ) {
     const { posts, totalPosts } = await this.postService.getPosts({
       take,