+++ apps/example-next/tests/react-query.spec.tsx
@@ -117,7 +117,7 @@ const postsRouter = c.router(
     baseHeaders: z.object({
       'x-test': z.string(),
     }),
-  }
+  },
 );
 
 // Three endpoints, two for posts, and one for health
@@ -212,7 +212,7 @@ describe('react-query', () => {
         }),
       {
         wrapper,
-      }
+      },
     );
 
     expect(result.current.data).toStrictEqual(undefined);
@@ -249,7 +249,7 @@ describe('react-query', () => {
         }),
       {
         wrapper,
-      }
+      },
     );
 
     expect(result.current.data).toStrictEqual(undefined);
@@ -284,7 +284,7 @@ describe('react-query', () => {
         }),
       {
         wrapper,
-      }
+      },
     );
 
     expect(result.current.data).toStrictEqual(undefined);
@@ -342,7 +342,7 @@ describe('react-query', () => {
 
     const { result } = renderHook(
       () => client.health.useQuery(['health'], {}, { retry: () => false }),
-      { wrapper }
+      { wrapper },
     );
 
     expect(result.current.data).toStrictEqual(undefined);
@@ -390,7 +390,7 @@ describe('react-query', () => {
           content: '',
           authorId: '1',
         },
-      })
+      }),
     );
 
     expect(api).toHaveBeenCalledWith({
@@ -447,7 +447,7 @@ describe('react-query', () => {
         }),
       {
         wrapper,
-      }
+      },
     );
 
     expect(result.current[0].data).toStrictEqual(undefined);
@@ -518,7 +518,7 @@ describe('react-query', () => {
         }),
       {
         wrapper,
-      }
+      },
     );
 
     expect(result.current[0].data).toStrictEqual(undefined);
@@ -593,7 +593,7 @@ describe('react-query', () => {
         }),
       {
         wrapper,
-      }
+      },
     );
 
     expect(result.current[0].data).toStrictEqual(undefined);
@@ -654,7 +654,7 @@ describe('react-query', () => {
       },
       {
         wrapper,
-      }
+      },
     );
 
     expect(api).toHaveBeenCalledWith({
@@ -683,7 +683,7 @@ describe('react-query', () => {
       },
       {
         wrapper,
-      }
+      },
     );
 
     expect(api).toHaveBeenCalledWith({
@@ -716,7 +716,7 @@ describe('react-query', () => {
       },
       {
         wrapper,
-      }
+      },
     );
 
     expect(result.current).resolves.toStrictEqual(ERROR_RESPONSE);
@@ -747,7 +747,7 @@ describe('react-query', () => {
       },
       {
         wrapper,
-      }
+      },
     );
 
     expect(api).toHaveBeenCalledWith({
@@ -777,7 +777,7 @@ describe('react-query', () => {
       },
       {
         wrapper,
-      }
+      },
     );
 
     await waitForNextUpdate();
@@ -789,7 +789,7 @@ describe('react-query', () => {
       },
       {
         wrapper,
-      }
+      },
     );
 
     expect(result.current).toStrictEqual(SUCCESS_RESPONSE);
@@ -837,11 +837,11 @@ describe('react-query', () => {
           },
           {
             staleTime: 10000,
-          }
+          },
         ),
       {
         wrapper,
-      }
+      },
     );
 
     await waitForNextUpdate();
@@ -853,7 +853,7 @@ describe('react-query', () => {
       },
       {
         wrapper,
-      }
+      },
     );
 
     const { result } = renderHook(
@@ -867,11 +867,11 @@ describe('react-query', () => {
           },
           {
             staleTime: 10000,
-          }
+          },
         ),
       {
         wrapper,
-      }
+      },
     );
 
     return waitFor(() => {