+++ libs/ts-rest/core/src/lib/fetch.spec.ts
@@ -100,7 +100,7 @@ describe('fetchApi', () => {
         },
         rawQuery: {},
         signal: undefined,
-      })
+      }),
     );
   });
 });