+++ libs/ts-rest/nest/src/lib/ts-rest-nest-handler.spec.ts
@@ -9,11 +9,15 @@ import { z } from 'zod';
 import {
   ArgumentsHost,
   Body,
+  CallHandler,
   Catch,
   Controller,
   ExceptionFilter,
+  ExecutionContext,
   Get,
   HttpException,
+  Injectable,
+  NestInterceptor,
   Post,
   UploadedFile,
   UseInterceptors,
@@ -25,9 +29,10 @@ import path = require('path');
 import { FileInterceptor } from '@nestjs/platform-express';
 import { FastifyAdapter } from '@nestjs/platform-fastify';
 import { Response } from 'express';
+import { map, Observable } from 'rxjs';
 
 export type Equal<a, b> = (<T>() => T extends a ? 1 : 2) extends <
-  T
+  T,
 >() => T extends b ? 1 : 2
   ? true
   : false;
@@ -52,7 +57,7 @@ describe('doesUrlMatchContractPath', () => {
     'should return $expected when contractPath is $contractPath and url is $url',
     ({ contractPath, url, expected }) => {
       expect(doesUrlMatchContractPath(contractPath, url)).toBe(expected);
-    }
+    },
   );
 });
 
@@ -467,6 +472,66 @@ describe('ts-rest-nest-handler', () => {
         expect(response.body).toEqual({ message: 'ok' });
       });
     });
+
+    it('should be able to intercept', async () => {
+      const c = initContract();
+
+      const contract = c.router({
+        getRequest: {
+          path: '/',
+          method: 'GET',
+          responses: {
+            200: z.object({
+              message: z.string(),
+            }),
+          },
+        },
+      });
+
+      @Injectable()
+      class TestInterceptor implements NestInterceptor {
+        intercept(
+          context: ExecutionContext,
+          next: CallHandler,
+        ): Observable<any> {
+          return next.handle().pipe(
+            map((data) => ({
+              message: 'intercepted',
+              oldMessage: data.message,
+            })),
+          );
+        }
+      }
+
+      @Controller()
+      class SingleHandlerTestController {
+        @TsRestHandler(contract)
+        @UseInterceptors(TestInterceptor)
+        async postRequest() {
+          return tsRestHandler(contract, {
+            getRequest: async () => ({
+              status: 200,
+              body: { message: 'ok' },
+            }),
+          });
+        }
+      }
+
+      const moduleRef = await Test.createTestingModule({
+        controllers: [SingleHandlerTestController],
+      }).compile();
+
+      const app = moduleRef.createNestApplication();
+      await app.init();
+
+      const response = await supertest(app.getHttpServer()).get('/').send();
+
+      expect(response.status).toBe(200);
+      expect(response.body).toEqual({
+        message: 'intercepted',
+        oldMessage: 'ok',
+      });
+    });
   });
 
   describe('single-handler api', () => {
@@ -671,7 +736,7 @@ describe('ts-rest-nest-handler', () => {
         @UseInterceptors(FileInterceptor('file'))
         nestMulti(
           @Body() body: { messageAsField: string },
-          @UploadedFile() file: File
+          @UploadedFile() file: File,
         ) {
           return {
             messageAsField: body.messageAsField,
@@ -786,7 +851,7 @@ describe('ts-rest-nest-handler', () => {
             200: z.array(
               z.object({
                 message: z.string(),
-              })
+              }),
             ),
           },
         },
@@ -970,7 +1035,7 @@ describe('ts-rest-nest-handler', () => {
         },
         {
           strictStatusCodes: true,
-        }
+        },
       );
 
       @Controller()
@@ -1185,7 +1250,7 @@ describe('ts-rest-nest-handler', () => {
               status: 200,
               body: { message: 'hello' },
             }),
-          }
+          },
         );
       }
     }
@@ -1232,7 +1297,7 @@ describe('ts-rest-nest-handler', () => {
           200: z.array(
             z.object({
               id: z.number(),
-            })
+            }),
           ),
         },
       },
@@ -1438,7 +1503,7 @@ describe('ts-rest-nest-handler', () => {
           200: z.array(
             z.object({
               id: z.number(),
-            })
+            }),
           ),
         },
       },
@@ -1446,15 +1511,19 @@ describe('ts-rest-nest-handler', () => {
 
     const cause = new Error('the root cause');
 
-    const error = new TsRestException(contract.getPosts, {
-      status: 400,
-      body: {
-        message: 'Something went wrong'
-      }
-    }, {
-      cause
-    })
+    const error = new TsRestException(
+      contract.getPosts,
+      {
+        status: 400,
+        body: {
+          message: 'Something went wrong',
+        },
+      },
+      {
+        cause,
+      },
+    );
 
     expect(error.cause).toEqual(cause);
-  })
+  });
 });