+++ libs/ts-rest/open-api/src/lib/ts-rest-open-api.spec.ts
@@ -1,6 +1,7 @@
 import { initContract } from '@ts-rest/core';
 import { z } from 'zod';
 import { generateOpenApi } from './ts-rest-open-api';
+import { extendApi } from '@anatine/zod-openapi';
 
 const c = initContract();
 
@@ -68,7 +69,7 @@ const postsRouter = c.router({
     method: 'GET',
     path: `/posts/:id/comments/:commentId`,
     pathParams: z.object({
-      commentId: z.string().length(5).describe("the comment ID"),
+      commentId: z.string().length(5).describe('the comment ID'),
     }),
     responses: {
       200: c.type<Post | null>(),
@@ -101,6 +102,35 @@ const router = c.router({
       200: c.type<{ message: string }>(),
     },
   },
+  mediaExamples: {
+    method: 'POST',
+    path: '/media-examples',
+    query: z.object({
+      foo: extendApi(z.string(), {
+        // this will only be added when jsonQuery is enabled
+        mediaExamples: {
+          one: { value: 'foo' },
+          two: { value: 'bar' },
+        },
+      }),
+    }),
+    summary: 'Examples API',
+    description: `Check that examples can be added to body and response types`,
+    body: extendApi(z.object({ id: z.string() }), {
+      mediaExamples: {
+        one: { value: { id: 'foo' } },
+        two: { value: { id: 'bar' } },
+      },
+    }),
+    responses: {
+      200: extendApi(z.object({ id: z.string() }), {
+        mediaExamples: {
+          three: { value: { id: 'foo' } },
+          four: { value: { id: 'bar' } },
+        },
+      }),
+    },
+  },
 });
 
 const expectedApiDoc = {
@@ -121,6 +151,66 @@ const expectedApiDoc = {
           },
         },
         summary: 'Health API',
+        tags: [],
+      },
+    },
+    '/media-examples': {
+      post: {
+        deprecated: undefined,
+        description: `Check that examples can be added to body and response types`,
+        parameters: [
+          {
+            in: 'query',
+            name: 'foo',
+            required: true,
+            schema: {
+              type: 'string',
+            },
+          },
+        ],
+        requestBody: {
+          description: 'Body',
+          content: {
+            'application/json': {
+              schema: {
+                properties: {
+                  id: {
+                    type: 'string',
+                  },
+                },
+                required: ['id'],
+                type: 'object',
+              },
+              examples: {
+                one: { value: { id: 'foo' } },
+                two: { value: { id: 'bar' } },
+              },
+            },
+          },
+        },
+        responses: {
+          '200': {
+            content: {
+              'application/json': {
+                examples: {
+                  three: { value: { id: 'foo' } },
+                  four: { value: { id: 'bar' } },
+                },
+                schema: {
+                  properties: {
+                    id: {
+                      type: 'string',
+                    },
+                  },
+                  required: ['id'],
+                  type: 'object',
+                },
+              },
+            },
+            description: `200`,
+          },
+        },
+        summary: 'Examples API',
         tags: [],
       },
     },
@@ -408,7 +498,7 @@ describe('ts-rest-open-api', () => {
         {
           info: { title: 'Blog API', version: '0.1' },
         },
-        { setOperationId: true }
+        { setOperationId: true },
       );
 
       expect(apiDoc).toEqual({
@@ -418,6 +508,12 @@ describe('ts-rest-open-api', () => {
             get: {
               ...expectedApiDoc.paths['/health'].get,
               operationId: 'health',
+            },
+          },
+          '/media-examples': {
+            post: {
+              ...expectedApiDoc.paths['/media-examples'].post,
+              operationId: 'mediaExamples',
             },
           },
           '/posts': {
@@ -464,13 +560,41 @@ describe('ts-rest-open-api', () => {
         {
           info: { title: 'Blog API', version: '0.1' },
         },
-        { jsonQuery: true }
+        { jsonQuery: true },
       );
 
       expect(apiDoc).toEqual({
         ...expectedApiDoc,
         paths: {
           ...expectedApiDoc.paths,
+          '/media-examples': {
+            ...expectedApiDoc.paths['/media-examples'],
+            post: {
+              ...expectedApiDoc.paths['/media-examples'].post,
+              parameters: [
+                {
+                  content: {
+                    'application/json': {
+                      examples: {
+                        one: {
+                          value: 'foo',
+                        },
+                        two: {
+                          value: 'bar',
+                        },
+                      },
+                      schema: {
+                        type: 'string',
+                      },
+                    },
+                  },
+                  in: 'query',
+                  name: 'foo',
+                  required: true,
+                },
+              ],
+            },
+          },
           '/posts': {
             ...expectedApiDoc.paths['/posts'],
             get: {
@@ -557,8 +681,8 @@ describe('ts-rest-open-api', () => {
           {
             info: { title: 'Blog API', version: '0.1' },
           },
-          { setOperationId: true }
-        )
+          { setOperationId: true },
+        ),
       ).toThrowError(/getPost/);
     });
 
@@ -572,7 +696,7 @@ describe('ts-rest-open-api', () => {
           },
           query: z
             .object({
-              foo: z.string().describe("Foo"),
+              foo: z.string().describe('Foo'),
             })
             .refine((v) => v.foo === 'bar', {
               message: 'foo must be bar',
@@ -604,6 +728,75 @@ describe('ts-rest-open-api', () => {
                   schema: {
                     type: 'string',
                   },
+                },
+              ],
+              responses: {
+                '200': {
+                  description: '200',
+                },
+              },
+              summary: undefined,
+              tags: [],
+            },
+          },
+        },
+      });
+    });
+
+    it('works with zod optional query parameters', () => {
+      const routerWithRefine = c.router({
+        endpointWithZodRefine: {
+          method: 'GET',
+          path: '/optional',
+          responses: {
+            200: c.type<null>(),
+          },
+          query: z.object({
+            foo: z
+              .object({
+                baz: z.string().describe('Baz').optional(),
+                bar: z.string().describe('Bar').optional(),
+              })
+              .describe('Foo')
+              .optional(),
+          }),
+        },
+      });
+
+      const schema = generateOpenApi(routerWithRefine, {
+        info: { title: 'Blog API', version: '0.1' },
+      });
+
+      expect(schema).toEqual({
+        info: {
+          title: 'Blog API',
+          version: '0.1',
+        },
+        openapi: '3.0.2',
+        paths: {
+          '/optional': {
+            get: {
+              deprecated: undefined,
+              description: undefined,
+              parameters: [
+                {
+                  description: 'Foo',
+                  in: 'query',
+                  name: 'foo',
+                  schema: {
+                    properties: {
+                      bar: {
+                        description: 'Bar',
+                        type: 'string',
+                      },
+                      baz: {
+                        description: 'Baz',
+                        type: 'string',
+                      },
+                    },
+                    type: 'object',
+                  },
+                  style: 'deepObject',
                 },
               ],
               responses: {
@@ -706,7 +899,7 @@ describe('ts-rest-open-api', () => {
               parameters: [],
               requestBody: {
                 content: {
-                  "multipart/form-data": {
+                  'multipart/form-data': {
                     schema: {
                       properties: {
                         file: {
@@ -718,7 +911,7 @@ describe('ts-rest-open-api', () => {
                     },
                   },
                 },
-                description: "Body",
+                description: 'Body',
               },
               responses: {
                 '200': {