+++ libs/ts-rest/serverless/src/index.ts
@@ -1,0 +1,4 @@
+export * from './lib/http-error';
+export * from './lib/request';
+export * from './lib/response';
+export { RequestValidationError, ResponseValidationError } from './lib/types';