+++ libs/ts-rest/nest/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ts-rest/nest
 
+## 3.35.1
+
+### Patch Changes
+
+- b3fde9b: Fix Nest.js interceptor returninng promise when using new Nest.js handlers
+- 14f9416: Allow Nest.js handlers to return non-promise responses
+
+## 3.35.0
+
+## 3.34.0
+
+## 3.33.1
+
 ## 3.33.0
 
 ## 3.32.0